exo 6

+++ app/Http/Controllers/ProductController.php
@@ -4,12 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+
     public function index()
     {
         $products = Product::with('user')->get();
@@ -17,26 +16,19 @@ class ProductController extends Controller
         return view('products.index', compact('products'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
         return view('products.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
+
     public function store(Request $request)
     {
-        // Validation basique
         $validated = $request->validate([
             'name'  => ['required', 'string', 'max:255'],
             'price' => ['required', 'numeric', 'min:0'],
         ]);
 
-        // Création du produit pour l'utilisateur connecté
         Product::create([
             'name'     => $validated['name'],
             'price'    => $validated['price'],
@@ -49,27 +41,27 @@ class ProductController extends Controller
             ->with('status', 'Produit créé avec succès.');
     }
 
-    /**
-     * Display the specified resource.
-     */
+
     public function show(Product $product)
     {
+        Gate::authorize('view-product', $product);
+
         return view('products.show', compact('product'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
+
     public function edit(Product $product)
     {
+        Gate::authorize('manage-product', $product);
+
         return view('products.edit', compact('product'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
+ 
     public function update(Request $request,  Product $product)
     {
+        Gate::authorize('manage-product', $product);
+
         $validated = $request->validate([
             'name'  => ['required', 'string', 'max:255'],
             'price' => ['required', 'numeric', 'min:0'],
@@ -86,11 +78,11 @@ class ProductController extends Controller
             ->with('status', 'Produit mis à jour avec succès.');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
+
     public function destroy(Product $product)
     {
+        Gate::authorize('manage-product', $product);
+
         $product->delete();
 
         return redirect()
